[FEAT] FriendList Component 제작 및 CSS 수정
feat: #10
