+++ src/shared/component/ParticipateStudy/index.tsx
@@ -3,7 +3,7 @@ import PersonalList from "@/shared/component/ParticipateStudy/PersonalList";
 
 export default function ParticipateStudy() {
   return (
-    <div className="bg-white border-0 rounded-lg w-[420px] h-[465px]">
+    <div className="bg-white rounded-lg w-[420px] h-[465px]">
       <div className="font-bold text-[15px] mt-[16px] ml-[25px]">
         내가 참여중인 스터디
       </div>
